requirements.txt
tflite-runtime==2.14.0
numpy==1.24.3
opencv-python-headless==4.8.0.74
PILLOW==9.4.0